requirements.txt
discord.py
discord.py[voice] 
timeago 
dnspython 
aiohttp 
youtube_dl 
ffmpeg 
pymongo
psutil
discord_webhook
bs4
requests
clashroyale
motor
lxml
tweepy
oauthlib
pyyaml
pynite
Pillow
lavalink
brawlstars
aioconsole
cachetools
numpy
better-profanity
mysql-connector-python
flask
flask-discord